multidispatch: fixed type_index
using meta::filter to ignore non-handles leads to empty lists if
no handles are left in the listi while processing the list. This
causes meta::cartesian_product to fail, since cartesian_product<>
is not defined. Using meta::make_list yields the same results, but
covers the cases where the tail of the given list are all non-handles.